refactor: Remove legacy PrefillDecode routing mode
Consolidate prefill-decode routing by removing the old PrefillDecode
mode in favor of VllmPrefillDecode. This simplifies the codebase by
having a single, vLLM-optimized implementation for P/D disaggregation.

- Remove PrefillDecode variant from RoutingMode enum
- Delete src/routers/http/pd_router.rs and src/routers/grpc/pd_router.rs
- Update tests to use VllmPrefillDecode
- Simplify router factory and validation logic
