+++ src/routers/grpc/mod.rs
@@ -1,4 +1,3 @@
 //! gRPC router implementations
 
-pub mod pd_router;
 pub mod router;
