+++ src/routers/http/mod.rs
@@ -3,7 +3,6 @@
 pub mod dp_utils;
 pub mod logprobs_merge;
 pub mod openai_router;
-pub mod pd_router;
 pub mod pd_types;
 pub mod router;
 pub mod vllm_pd_router;
